evdev

Struct AttributeSetRef

pub struct AttributeSetRef<T> { /* private fields */ }

A collection of bits representing either device capability or state.

This can be used to iterate across all keys supported by a keyboard, or all buttons supported by a joystick. You can also query directly whether a specific bit is set (corresponding to whether a key or button is depressed).

Implementations

impl<T: EvdevEnum> AttributeSetRef<T>

pub fn contains(&self, attr: T) -> bool

Returns true if this AttributeSet contains the passed T.

pub fn iter(&self) -> AttributeSetRefIter<'_, T>

Provides an iterator over all “set” bits in the collection.

pub fn insert(&mut self, attr: T)

pub fn remove(&mut self, attr: T)