Struct evdev::Device

pub struct Device {
    // some fields omitted
}

Methods

impl Device

fn fd(&self) -> RawFd

fn events_supported(&self) -> Types

fn name(&self) -> &CString

fn physical_path(&self) -> &Option<CString>

fn unique_name(&self) -> &Option<CString>

fn input_id(&self) -> input_id

fn properties(&self) -> Props

fn driver_version(&self) -> (u8, u8, u8)

fn keys_supported(&self) -> &FixedBitSet

fn relative_axes_supported(&self) -> RelativeAxis

fn absolute_axes_supported(&self) -> AbsoluteAxis

fn switches_supported(&self) -> Switch

fn leds_supported(&self) -> Led

fn misc_properties(&self) -> Misc

fn repeats_supported(&self) -> Repeat

fn sounds_supported(&self) -> Sound

fn state(&self) -> &DeviceState

fn open(path: &AsRef<Path>) -> Result<Device, Error>

fn sync_state(&mut self) -> Result<(), Error>

Synchronize the Device state with the kernel device state.

If there is an error at any point, the state will not be synchronized completely.

fn events_no_sync(&mut self) -> Result<RawEvents, Error>

Exposes the raw evdev events without doing synchronization on SYN_DROPPED.

fn events(&mut self) -> Result<RawEvents, Error>

Exposes the raw evdev events, doing synchronization on SYN_DROPPED.

Will insert "fake" events